tein::foreign

Type Alias MethodFn

pub type MethodFn = fn(&mut dyn Any, &MethodContext, &[Value]) -> Result<Value>;

A method on a foreign type, callable from Scheme.

Receives a mutable reference to the object (downcast inside the body), a limited context for creating return values, and the remaining arguments.

example

let method: MethodFn = |obj, _ctx, args| {
    let counter = obj.downcast_mut::<Counter>().unwrap();
    counter.n += 1;
    Ok(Value::Integer(counter.n))
};